+++ daemon/command/daemon.go
@@ -29,6 +29,7 @@ import (
 	"github.com/docker/docker/daemon/internal/builder-next/exporter"
 	"github.com/docker/docker/daemon/internal/libcontainerd/supervisor"
 	"github.com/docker/docker/daemon/internal/otelutil"
+	"github.com/docker/docker/daemon/internal/rootless"
 	"github.com/docker/docker/daemon/listeners"
 	dopts "github.com/docker/docker/daemon/pkg/opts"
 	"github.com/docker/docker/daemon/pkg/plugin"
@@ -53,7 +54,6 @@ import (
 	"github.com/docker/docker/pkg/homedir"
 	"github.com/docker/docker/pkg/pidfile"
 	"github.com/docker/docker/pkg/plugingetter"
-	"github.com/docker/docker/pkg/rootless"
 	"github.com/docker/go-connections/tlsconfig"
 	"github.com/moby/buildkit/session"
 	"github.com/moby/buildkit/util/tracing/detect"
