+++ daemon/config/config_linux.go
@@ -11,10 +11,10 @@ import (
 
 	"github.com/containerd/cgroups/v3"
 	"github.com/containerd/log"
+	"github.com/docker/docker/daemon/internal/rootless"
 	"github.com/docker/docker/daemon/libnetwork/drivers/bridge"
 	"github.com/docker/docker/daemon/pkg/opts"
 	"github.com/docker/docker/pkg/homedir"
-	"github.com/docker/docker/pkg/rootless"
 	"github.com/moby/moby/api/types/container"
 	"github.com/moby/moby/api/types/system"
 	"github.com/pkg/errors"
