+++ pkg/plugins/discovery_unix.go
@@ -3,10 +3,10 @@
 package plugins
 
 import (
+	"os"
 	"path/filepath"
 
 	"github.com/docker/docker/pkg/homedir"
-	"github.com/docker/docker/pkg/rootless"
 )
 
 func rootlessConfigPluginsPath() string {
@@ -25,7 +25,8 @@ func rootlessLibPluginsPath() string {
 
 // specsPaths is the non-Windows implementation of [SpecsPaths].
 func specsPaths() []string {
-	if rootless.RunningWithRootlessKit() {
+	// TODO(thaJeztah): switch back to daemon/internal/rootless.RunningWithRootlessKit if this package moves internal to the daemon.
+	if os.Getenv("ROOTLESSKIT_STATE_DIR") != "" {
 		return []string{rootlessConfigPluginsPath(), rootlessLibPluginsPath()}
 	}
 	return []string{"/etc/docker/plugins", "/usr/lib/docker/plugins"}
